dom/base/test/test_copypaste.xul

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=888839
-->
<window title="Mozilla Bug 888839"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>

  <script type="application/javascript"><![CDATA[

SimpleTest.waitForExplicitFinish();
addLoadEvent(runTest);

function runTest() {
  let desc = document.querySelector("description");
  window.getSelection().selectAllChildren(desc);

  let expected = "\n    hello\n    world\n  ";
  SimpleTest.waitForClipboard(expected, function() {
    window.docShell
          .contentViewer
          .QueryInterface(Ci.nsIContentViewerEdit)
          .copySelection();
  }, function() {
    ok(true, "Paste is not HTML, so it should not be pretty printed");
    SimpleTest.finish();
  });
}

  ]]></script>

  <description style="-moz-user-focus: normal; -moz-user-select: text;"><![CDATA[
    hello
    world
  ]]></description>

  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=888839"
     target="_blank">Mozilla Bug 888839</a>
  </body>
</window>
